rag/requirements.txt
langchain-huggingface
python-dotenv
langchain-community
rank-bm25
langchain_together
together
lime
pandas
langchain_ollama
nltk
pinecone
faiss-cpu
structlog
